orchestration/base-domain/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "base-domain"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["orchestration", "callbacks", "base_domain"]
authors = [
  { name = "alek-ellegard", email = "alek-ellegard@caesari.se" },
]
dependencies = [
    "uvloop>=0.21.0",
    "orjson>=3.10.15",
    "pyzmq>=27.1.0",
    "asyncio>=4.0.0",
    "pyzmq>=27.1.0",
]

[dependency-groups]  # modern uv pattern for dev deps
dev = [
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "ruff>=0.13.3",
]

[project.scripts]
run-base-domain = "base_domain.entrypoint:run"

[tool.hatch.envs.types]
dependencies = ["mypy>=1.14.1"]

[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src tests}"

[tool.hatch.envs.dev]
installer = "uv"
dependencies = [
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.13.3",
]


[tool.coverage.run]
source_pkgs = ["base_domain", "tests"]
branch = true
parallel = true

[tool.coverage.paths]
base_domain = ["src/base_domain", "*/base-domain/src/base_domain"]
tests = ["tests", "*/base-domain/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
